requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.0.55
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-openrouter>=0.0.3

# Streamlit for UI
streamlit>=1.29.0

# Environment and utilities
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# Optional: For enhanced functionality
# langchain-community>=0.0.20  # For additional integrations
# pydantic>=2.0.0  # For data validation

# Development dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0